move legacy launch to point of use
