--- pkg/cmd/server/etcd/doc.go
@@ -1,2 +1,0 @@
-// Package etcd provides objects for creating an embedded etcd server process.
-package etcd
